requirements.txt
flask
python-dotenv
ipykernel
torch
pillow
transformers
openai
diffusers
accelerate